server/requirements.txt
bottle
pytest-approvaltests